Added the dashboard backend logic for user information regarding view dashboard , view profile and update the profile
